+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/internal/template/parser/TokensTransformerListener.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.journal.internal.transformer;
+package com.liferay.journal.internal.template.parser;
 
 import com.liferay.journal.constants.JournalPortletKeys;
 import com.liferay.journal.constants.JournalTransformerListenerKeys;
